# arka-meta/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Arka R1 Backend API (M2)
  version: 0.1.0
  description: >
    Contrats initiaux pour M2 (SSE chat, Documents, Agent Events, Export mémoire, Metrics).
servers:
  - url: https://api.arka.local
security:
  - bearerAuth: []
paths:
...
  /api/documents:
    get:
      summary: List project documents
      operationId: listDocuments
      security: [{ bearerAuth: [] }]
      parameters:
        - name: project
          in: query
          required: false
          schema: { type: string, example: arka }
      responses:
        '200':
          description: Documents
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Document' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      summary: Upload a document
      operationId: uploadDocument
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                project:
                  type: string
                  example: arka
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Document' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
      required: [error]
...
    Document:
      type: object
      properties:
        id: { type: integer }
        project_id: { type: string }
        name: { type: string }
        mime: { type: string }
        size: { type: integer, minimum: 0 }
        storage_key: { type: string }
        tags:
          type: array
          items: { type: string }
        created_at: { type: string, format: date-time }
      required: [id, name, mime, size, created_at]
...